+++ cvat/apps/dataset_manager/tests/test_formats.py
@@ -27,7 +27,7 @@ from cvat.apps.dataset_manager.util import make_zip_archive
 from cvat.apps.engine.models import Task
 
 
-def generate_image_file(filename, size=(100, 50)):
+def generate_image_file(filename, size=(100, 100)):
     f = BytesIO()
     image = Image.new('RGB', size=size)
     image.save(f, 'jpeg')
@@ -297,6 +297,8 @@ class TaskExportTest(_DbTestBase):
             'Kitti Raw Format 1.0',
             'Sly Point Cloud Format 1.0',
             'LFW 1.0',
+            'Cityscapes 1.0',
+            'Open Images V6 1.0'
         })
 
     def test_import_formats_query(self):
@@ -324,6 +326,8 @@ class TaskExportTest(_DbTestBase):
             'Kitti Raw Format 1.0',
             'Sly Point Cloud Format 1.0',
             'LFW 1.0',
+            'Cityscapes 1.0',
+            'Open Images V6 1.0',
             'Datumaro 1.0',
             'Datumaro 3D 1.0',
         })
@@ -372,7 +376,8 @@ class TaskExportTest(_DbTestBase):
             ('ICDAR Recognition 1.0', 'icdar_word_recognition'),
             ('ICDAR Localization 1.0', 'icdar_text_localization'),
             ('ICDAR Segmentation 1.0', 'icdar_text_segmentation'),
-            ('LFW 1.0', 'lfw')
+            ('LFW 1.0', 'lfw'),
+            # ('Cityscapes 1.0', 'cityscapes'), does not support, empty annotations
         ]:
             with self.subTest(format=format_name):
                 if not dm.formats.registry.EXPORT_FORMATS[format_name].ENABLED:
@@ -383,16 +388,6 @@ class TaskExportTest(_DbTestBase):
 
                 def check(file_path):
                     def load_dataset(src):
-                        if importer_name == 'datumaro_project':
-                            project = datumaro.components.project. \
-                                Project.load(src)
-
-                            # NOTE: can't import cvat.utils.cli
-                            # for whatever reason, so remove the dependency
-                            #
-                            project.config.remove('sources')
-
-                            return project.make_dataset()
                         return datumaro.components.dataset. \
                             Dataset.import_from(src, importer_name, env=dm_env)
 
@@ -703,6 +698,10 @@ class TaskAnnotationsImportTest(_DbTestBase):
                         }
                     ]
                 },
+                {
+                    "name": "background",
+                    "attributes": [],
+                },
                 {"name": "person"}
             ]
 
@@ -839,7 +838,7 @@ class TaskAnnotationsImportTest(_DbTestBase):
                     {
                         "frame": 0,
                         "attributes": [],
-                        "points": [1.0, 2.1, 10.6, 53.22, 100, 300.222],
+                        "points": [1.0, 2.1, 10.6, 53.22, 90, 90.222],
                         "type": "polygon",
                         "occluded": False,
                         "outside": False
@@ -880,10 +879,10 @@ class TaskAnnotationsImportTest(_DbTestBase):
             elif annotation_format == "MOTS PNG 1.0":
                 tracks = [track_wo_attrs]
             else:
-                shapes = [rectangle_shape_wo_attrs,
+                shapes = [rectangle_shape_wo_attrs, \
                     rectangle_shape_with_attrs]
-                tags = tag_wo_attrs
-                tracks = track_wo_attrs
+                tags = [tag_wo_attrs]
+                tracks = [track_wo_attrs]
 
         annotations = {
             "version": 0,